[C] Move clock code out of the aeronmd.h and aeron_driver.c into a separate compilation unit.
